requirements.txt

gspread
google-auth
selenium
persiantools
google-api-python-client
python-telegram-bot==20.7
